+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed `StopWatch` to `Stopwatch` and `StopWatchError` to `StopwatchError` to better reflect that "stopwatch" is a single English word
+
 ## [1.0.0] - 2025-10-19
 
 ### Added